refactor(frontend): extract history list item into reusable component
